+++ IMS.CoreBusiness/Product.cs
@@ -15,7 +15,7 @@ namespace IMS.CoreBusiness
         public int Quantity { get; set; }
 
         [Range(0, int.MaxValue, ErrorMessage = "单价必须大于等于0")]
-        public double Price { get; set; }
+        public Decimal Price { get; set; }
 
         [Product_EnsurePriceIsGreaterThanInventoriesCost]
         public List<ProductInventory> ProductInventories { get; set; } = new List<ProductInventory>();
