+++ IMS.UseCases/PluginInterfaces/IInventoryTransactionRepository.cs
@@ -4,7 +4,7 @@ namespace IMS.UseCases.PluginInterfaces
 {
     public interface IInventoryTransactionRepository
     {
-        Task PurchaseAsync(string orderNumber, Inventory inventory, int quantity, string doneBy, double price);
-        Task ProduceAsync(string productionNumber, Inventory inventory, int quantityToConsume, string doneBy, double price);
+        Task PurchaseAsync(string orderNumber, Inventory inventory, int quantity, string doneBy, decimal price);
+        Task ProduceAsync(string productionNumber, Inventory inventory, int quantityToConsume, string doneBy, decimal price);
     }
 }
